Fix overflow in toasting time

+++ toast.h
@@ -8,13 +8,13 @@
 namespace toasting::toast {
 
   struct Toast {
-    uint16_t time;
+    uint32_t time;
     const char* name;
-    void (&render)(int16_t,int16_t);
+    void (&render)(int16_t, int16_t);
   };
 
-  void renderBread(int16_t x,int16_t y);
-  void renderBagel(int16_t x,int16_t y);
+  void renderBread(int16_t x, int16_t y);
+  void renderBagel(int16_t x, int16_t y);
 
   constexpr
   std_array<std_array<Toast, 2>, 1> toasts
@@ -28,8 +28,8 @@ namespace toasting::toast {
 
   extern ToasterWattage curWattage;
 
-  constexpr uint16_t conversion[3] = {1400, 1200, 1000};
-  constexpr uint16_t convert(ToasterWattage w)
+  constexpr uint32_t conversion[3] = {1400, 1200, 1000};
+  constexpr uint32_t convert(ToasterWattage w)
   { return conversion[static_cast<uint8_t>(w)]; }
 
   void setup();
